Struct aws_sdk_clouddirectory::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonCloudDirectory_20170111.

This client allows ergonomic access to a AmazonCloudDirectory_20170111-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_facet_to_object(&self) -> AddFacetToObject<C, M, R>

pub fn apply_schema(&self) -> ApplySchema<C, M, R>

pub fn attach_object(&self) -> AttachObject<C, M, R>

pub fn attach_policy(&self) -> AttachPolicy<C, M, R>

pub fn attach_to_index(&self) -> AttachToIndex<C, M, R>

pub fn attach_typed_link(&self) -> AttachTypedLink<C, M, R>

pub fn batch_read(&self) -> BatchRead<C, M, R>

pub fn batch_write(&self) -> BatchWrite<C, M, R>

pub fn create_directory(&self) -> CreateDirectory<C, M, R>

pub fn create_facet(&self) -> CreateFacet<C, M, R>

pub fn create_index(&self) -> CreateIndex<C, M, R>

pub fn create_object(&self) -> CreateObject<C, M, R>

pub fn create_schema(&self) -> CreateSchema<C, M, R>

pub fn create_typed_link_facet(&self) -> CreateTypedLinkFacet<C, M, R>

pub fn delete_directory(&self) -> DeleteDirectory<C, M, R>

pub fn delete_facet(&self) -> DeleteFacet<C, M, R>

pub fn delete_object(&self) -> DeleteObject<C, M, R>

pub fn delete_schema(&self) -> DeleteSchema<C, M, R>

pub fn delete_typed_link_facet(&self) -> DeleteTypedLinkFacet<C, M, R>

pub fn detach_from_index(&self) -> DetachFromIndex<C, M, R>

pub fn detach_object(&self) -> DetachObject<C, M, R>

pub fn detach_policy(&self) -> DetachPolicy<C, M, R>

pub fn detach_typed_link(&self) -> DetachTypedLink<C, M, R>

pub fn disable_directory(&self) -> DisableDirectory<C, M, R>

pub fn enable_directory(&self) -> EnableDirectory<C, M, R>

pub fn get_applied_schema_version(&self) -> GetAppliedSchemaVersion<C, M, R>

pub fn get_directory(&self) -> GetDirectory<C, M, R>

pub fn get_facet(&self) -> GetFacet<C, M, R>

pub fn get_link_attributes(&self) -> GetLinkAttributes<C, M, R>

pub fn get_object_attributes(&self) -> GetObjectAttributes<C, M, R>

pub fn get_object_information(&self) -> GetObjectInformation<C, M, R>

pub fn get_schema_as_json(&self) -> GetSchemaAsJson<C, M, R>

pub fn get_typed_link_facet_information(
&self
) -> GetTypedLinkFacetInformation<C, M, R>

pub fn list_applied_schema_arns(&self) -> ListAppliedSchemaArns<C, M, R>

pub fn list_attached_indices(&self) -> ListAttachedIndices<C, M, R>

pub fn list_development_schema_arns(&self) -> ListDevelopmentSchemaArns<C, M, R>

pub fn list_directories(&self) -> ListDirectories<C, M, R>

pub fn list_facet_attributes(&self) -> ListFacetAttributes<C, M, R>

pub fn list_facet_names(&self) -> ListFacetNames<C, M, R>

pub fn list_incoming_typed_links(&self) -> ListIncomingTypedLinks<C, M, R>

pub fn list_index(&self) -> ListIndex<C, M, R>

pub fn list_managed_schema_arns(&self) -> ListManagedSchemaArns<C, M, R>

pub fn list_object_attributes(&self) -> ListObjectAttributes<C, M, R>

pub fn list_object_children(&self) -> ListObjectChildren<C, M, R>

pub fn list_object_parent_paths(&self) -> ListObjectParentPaths<C, M, R>

pub fn list_object_parents(&self) -> ListObjectParents<C, M, R>

pub fn list_object_policies(&self) -> ListObjectPolicies<C, M, R>

pub fn list_outgoing_typed_links(&self) -> ListOutgoingTypedLinks<C, M, R>

pub fn list_policy_attachments(&self) -> ListPolicyAttachments<C, M, R>

pub fn list_published_schema_arns(&self) -> ListPublishedSchemaArns<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_typed_link_facet_attributes(
&self
) -> ListTypedLinkFacetAttributes<C, M, R>

pub fn list_typed_link_facet_names(&self) -> ListTypedLinkFacetNames<C, M, R>

pub fn lookup_policy(&self) -> LookupPolicy<C, M, R>

pub fn publish_schema(&self) -> PublishSchema<C, M, R>

pub fn put_schema_from_json(&self) -> PutSchemaFromJson<C, M, R>

pub fn remove_facet_from_object(&self) -> RemoveFacetFromObject<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_facet(&self) -> UpdateFacet<C, M, R>

pub fn update_link_attributes(&self) -> UpdateLinkAttributes<C, M, R>

pub fn update_object_attributes(&self) -> UpdateObjectAttributes<C, M, R>

pub fn update_schema(&self) -> UpdateSchema<C, M, R>

pub fn update_typed_link_facet(&self) -> UpdateTypedLinkFacet<C, M, R>

pub fn upgrade_applied_schema(&self) -> UpgradeAppliedSchema<C, M, R>

pub fn upgrade_published_schema(&self) -> UpgradePublishedSchema<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self