Module aws_sdk_clouddirectory::error

Modules

access_denied_exception

See AccessDeniedException

batch_write_exception

See BatchWriteException

cannot_list_parent_of_root_exception

See CannotListParentOfRootException

directory_already_exists_exception

See DirectoryAlreadyExistsException

directory_deleted_exception

See DirectoryDeletedException

directory_not_disabled_exception

See DirectoryNotDisabledException

directory_not_enabled_exception

See DirectoryNotEnabledException

facet_already_exists_exception

See FacetAlreadyExistsException

facet_in_use_exception

See FacetInUseException

facet_not_found_exception

See FacetNotFoundException

facet_validation_exception

See FacetValidationException

incompatible_schema_exception

See IncompatibleSchemaException

indexed_attribute_missing_exception

See IndexedAttributeMissingException

internal_service_exception

See InternalServiceException

invalid_arn_exception

See InvalidArnException

invalid_attachment_exception

See InvalidAttachmentException

invalid_facet_update_exception

See InvalidFacetUpdateException

invalid_next_token_exception

See InvalidNextTokenException

invalid_rule_exception

See InvalidRuleException

invalid_schema_doc_exception

See InvalidSchemaDocException

invalid_tagging_request_exception

See InvalidTaggingRequestException

limit_exceeded_exception

See LimitExceededException

link_name_already_in_use_exception

See LinkNameAlreadyInUseException

not_index_exception

See NotIndexException

not_node_exception

See NotNodeException

not_policy_exception

See NotPolicyException

object_already_detached_exception

See ObjectAlreadyDetachedException

object_not_detached_exception

See ObjectNotDetachedException

resource_not_found_exception

See ResourceNotFoundException

retryable_conflict_exception

See RetryableConflictException

schema_already_exists_exception

See SchemaAlreadyExistsException

schema_already_published_exception

See SchemaAlreadyPublishedException

still_contains_links_exception

See StillContainsLinksException

unsupported_index_type_exception

See UnsupportedIndexTypeException

validation_exception

See ValidationException

Structs

AccessDeniedException

Access denied or directory not found. Either you don't have permissions for this directory or the directory does not exist. Try calling ListDirectories and check your permissions.

AddFacetToObjectError

ApplySchemaError

AttachObjectError

AttachPolicyError

AttachToIndexError

AttachTypedLinkError

BatchReadError

BatchWriteError

BatchWriteException

A BatchWrite exception has occurred.

CannotListParentOfRootException

Cannot list the parents of a Directory root.

CreateDirectoryError

CreateFacetError

CreateIndexError

CreateObjectError

CreateSchemaError

CreateTypedLinkFacetError

DeleteDirectoryError

DeleteFacetError

DeleteObjectError

DeleteSchemaError

DeleteTypedLinkFacetError

DetachFromIndexError

DetachObjectError

DetachPolicyError

DetachTypedLinkError

DirectoryAlreadyExistsException

Indicates that a Directory could not be created due to a naming

DirectoryDeletedException

A directory that has been deleted and to which access has been attempted. Note: The

DirectoryNotDisabledException

An operation can only operate on a disabled directory.

DirectoryNotEnabledException

Operations are only permitted on enabled directories.

DisableDirectoryError

EnableDirectoryError

FacetAlreadyExistsException

A facet with the same name already exists.

FacetInUseException

Occurs when deleting a facet that contains an attribute that is a target to an

FacetNotFoundException

The specified Facet could not be found.

FacetValidationException

The Facet that you provided was not well formed or could not be

GetAppliedSchemaVersionError

GetDirectoryError

GetFacetError

GetLinkAttributesError

GetObjectAttributesError

GetObjectInformationError

GetSchemaAsJsonError

GetTypedLinkFacetInformationError

IncompatibleSchemaException

Indicates a failure occurred while performing a check for backward compatibility between the specified schema and the schema that is currently applied to the directory.

IndexedAttributeMissingException

An object has been attempted to be attached to an object that does not have the appropriate attribute value.

InternalServiceException

Indicates a problem that must be resolved by Amazon Web Services. This might be a transient error in which case you can retry your request until it succeeds. Otherwise, go to the AWS Service Health Dashboard site to see if there are any operational issues with the service.

InvalidArnException

Indicates that the provided ARN value is not valid.

InvalidAttachmentException

Indicates that an attempt to make an attachment was invalid. For example, attaching two nodes

InvalidFacetUpdateException

An attempt to modify a Facet resulted in an invalid schema

InvalidNextTokenException

Indicates that the NextToken value is not valid.

InvalidRuleException

Occurs when any of the rule parameter keys or values are invalid.

InvalidSchemaDocException

Indicates that the provided SchemaDoc value is not valid.

InvalidTaggingRequestException

Can occur for multiple reasons such as when you tag a resource that doesn’t exist or if you specify a higher number of tags for a resource than the allowed limit. Allowed limit is 50 tags per resource.

LimitExceededException

Indicates that limits are exceeded. See Limits for more information.

LinkNameAlreadyInUseException

Indicates that a link could not be created due to a naming conflict. Choose a different

ListAppliedSchemaArnsError

ListAttachedIndicesError

ListDevelopmentSchemaArnsError

ListDirectoriesError

ListFacetAttributesError

ListFacetNamesError

ListIncomingTypedLinksError

ListIndexError

ListManagedSchemaArnsError

ListObjectAttributesError

ListObjectChildrenError

ListObjectParentPathsError

ListObjectParentsError

ListObjectPoliciesError

ListOutgoingTypedLinksError

ListPolicyAttachmentsError

ListPublishedSchemaArnsError

ListTagsForResourceError

ListTypedLinkFacetAttributesError

ListTypedLinkFacetNamesError

LookupPolicyError

NotIndexException

Indicates that the requested operation can only operate on index objects.

NotNodeException

Occurs when any invalid operations are performed on an object that is not a node, such

NotPolicyException

Indicates that the requested operation can only operate on policy objects.

ObjectAlreadyDetachedException

Indicates that the object is not attached to the index.

ObjectNotDetachedException

Indicates that the requested operation cannot be completed because the object has not

PublishSchemaError

PutSchemaFromJsonError

RemoveFacetFromObjectError

ResourceNotFoundException

The specified resource could not be found.

RetryableConflictException

Occurs when a conflict with a previous successful write is detected. For example, if a write operation occurs on an object and then an attempt is made to read the object using “SERIALIZABLE” consistency, this exception may result. This generally occurs when the previous write did not have time to propagate to the host serving the current request. A retry (with appropriate backoff logic) is the recommended response to this exception.

SchemaAlreadyExistsException

Indicates that a schema could not be created due to a naming conflict. Please select a

SchemaAlreadyPublishedException

Indicates that a schema is already published.

StillContainsLinksException

The object could not be deleted because links still exist. Remove the links and then

TagResourceError

UnsupportedIndexTypeException

Indicates that the requested index type is not supported.

UntagResourceError

UpdateFacetError

UpdateLinkAttributesError

UpdateObjectAttributesError

UpdateSchemaError

UpdateTypedLinkFacetError

UpgradeAppliedSchemaError

UpgradePublishedSchemaError

ValidationException

Indicates that your request is malformed in some manner. See the exception

Enums

AddFacetToObjectErrorKind

ApplySchemaErrorKind

AttachObjectErrorKind

AttachPolicyErrorKind

AttachToIndexErrorKind

AttachTypedLinkErrorKind

BatchReadErrorKind

BatchWriteErrorKind

CreateDirectoryErrorKind

CreateFacetErrorKind

CreateIndexErrorKind

CreateObjectErrorKind

CreateSchemaErrorKind

CreateTypedLinkFacetErrorKind

DeleteDirectoryErrorKind

DeleteFacetErrorKind

DeleteObjectErrorKind

DeleteSchemaErrorKind

DeleteTypedLinkFacetErrorKind

DetachFromIndexErrorKind

DetachObjectErrorKind

DetachPolicyErrorKind

DetachTypedLinkErrorKind

DisableDirectoryErrorKind

EnableDirectoryErrorKind

GetAppliedSchemaVersionErrorKind

GetDirectoryErrorKind

GetFacetErrorKind

GetLinkAttributesErrorKind

GetObjectAttributesErrorKind

GetObjectInformationErrorKind

GetSchemaAsJsonErrorKind

GetTypedLinkFacetInformationErrorKind

ListAppliedSchemaArnsErrorKind

ListAttachedIndicesErrorKind

ListDevelopmentSchemaArnsErrorKind

ListDirectoriesErrorKind

ListFacetAttributesErrorKind

ListFacetNamesErrorKind

ListIncomingTypedLinksErrorKind

ListIndexErrorKind

ListManagedSchemaArnsErrorKind

ListObjectAttributesErrorKind

ListObjectChildrenErrorKind

ListObjectParentPathsErrorKind

ListObjectParentsErrorKind

ListObjectPoliciesErrorKind

ListOutgoingTypedLinksErrorKind

ListPolicyAttachmentsErrorKind

ListPublishedSchemaArnsErrorKind

ListTagsForResourceErrorKind

ListTypedLinkFacetAttributesErrorKind

ListTypedLinkFacetNamesErrorKind

LookupPolicyErrorKind

PublishSchemaErrorKind

PutSchemaFromJsonErrorKind

RemoveFacetFromObjectErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateFacetErrorKind

UpdateLinkAttributesErrorKind

UpdateObjectAttributesErrorKind

UpdateSchemaErrorKind

UpdateTypedLinkFacetErrorKind

UpgradeAppliedSchemaErrorKind

UpgradePublishedSchemaErrorKind