Module aws_sdk_clouddirectory::model

Modules

attribute_key

See AttributeKey

attribute_key_and_value

See AttributeKeyAndValue

attribute_name_and_value

See AttributeNameAndValue

batch_add_facet_to_object

See BatchAddFacetToObject

batch_add_facet_to_object_response

See BatchAddFacetToObjectResponse

batch_attach_object

See BatchAttachObject

batch_attach_object_response

See BatchAttachObjectResponse

batch_attach_policy

See BatchAttachPolicy

batch_attach_policy_response

See BatchAttachPolicyResponse

batch_attach_to_index

See BatchAttachToIndex

batch_attach_to_index_response

See BatchAttachToIndexResponse

batch_attach_typed_link

See BatchAttachTypedLink

batch_attach_typed_link_response

See BatchAttachTypedLinkResponse

batch_create_index

See BatchCreateIndex

batch_create_index_response

See BatchCreateIndexResponse

batch_create_object

See BatchCreateObject

batch_create_object_response

See BatchCreateObjectResponse

batch_delete_object

See BatchDeleteObject

batch_delete_object_response

See BatchDeleteObjectResponse

batch_detach_from_index

See BatchDetachFromIndex

batch_detach_from_index_response

See BatchDetachFromIndexResponse

batch_detach_object

See BatchDetachObject

batch_detach_object_response

See BatchDetachObjectResponse

batch_detach_policy

See BatchDetachPolicy

batch_detach_policy_response

See BatchDetachPolicyResponse

batch_detach_typed_link

See BatchDetachTypedLink

batch_detach_typed_link_response

See BatchDetachTypedLinkResponse

batch_get_link_attributes

See BatchGetLinkAttributes

batch_get_link_attributes_response

See BatchGetLinkAttributesResponse

batch_get_object_attributes

See BatchGetObjectAttributes

batch_get_object_attributes_response

See BatchGetObjectAttributesResponse

batch_get_object_information

See BatchGetObjectInformation

batch_get_object_information_response

See BatchGetObjectInformationResponse

batch_list_attached_indices

See BatchListAttachedIndices

batch_list_attached_indices_response

See BatchListAttachedIndicesResponse

batch_list_incoming_typed_links

See BatchListIncomingTypedLinks

batch_list_incoming_typed_links_response

See BatchListIncomingTypedLinksResponse

batch_list_index

See BatchListIndex

batch_list_index_response

See BatchListIndexResponse

batch_list_object_attributes

See BatchListObjectAttributes

batch_list_object_attributes_response

See BatchListObjectAttributesResponse

batch_list_object_children

See BatchListObjectChildren

batch_list_object_children_response

See BatchListObjectChildrenResponse

batch_list_object_parent_paths

See BatchListObjectParentPaths

batch_list_object_parent_paths_response

See BatchListObjectParentPathsResponse

batch_list_object_parents

See BatchListObjectParents

batch_list_object_parents_response

See BatchListObjectParentsResponse

batch_list_object_policies

See BatchListObjectPolicies

batch_list_object_policies_response

See BatchListObjectPoliciesResponse

batch_list_outgoing_typed_links

See BatchListOutgoingTypedLinks

batch_list_outgoing_typed_links_response

See BatchListOutgoingTypedLinksResponse

batch_list_policy_attachments

See BatchListPolicyAttachments

batch_list_policy_attachments_response

See BatchListPolicyAttachmentsResponse

batch_lookup_policy

See BatchLookupPolicy

batch_lookup_policy_response

See BatchLookupPolicyResponse

batch_read_exception

See BatchReadException

batch_read_operation

See BatchReadOperation

batch_read_operation_response

See BatchReadOperationResponse

batch_read_successful_response

See BatchReadSuccessfulResponse

batch_remove_facet_from_object

See BatchRemoveFacetFromObject

batch_remove_facet_from_object_response

See BatchRemoveFacetFromObjectResponse

batch_update_link_attributes

See BatchUpdateLinkAttributes

batch_update_link_attributes_response

See BatchUpdateLinkAttributesResponse

batch_update_object_attributes

See BatchUpdateObjectAttributes

batch_update_object_attributes_response

See BatchUpdateObjectAttributesResponse

batch_write_operation

See BatchWriteOperation

batch_write_operation_response

See BatchWriteOperationResponse

directory

See Directory

facet

See Facet

facet_attribute

See FacetAttribute

facet_attribute_definition

See FacetAttributeDefinition

facet_attribute_reference

See FacetAttributeReference

facet_attribute_update

See FacetAttributeUpdate

index_attachment

See IndexAttachment

link_attribute_action

See LinkAttributeAction

link_attribute_update

See LinkAttributeUpdate

object_attribute_action

See ObjectAttributeAction

object_attribute_range

See ObjectAttributeRange

object_attribute_update

See ObjectAttributeUpdate

object_identifier_and_link_name_tuple

See ObjectIdentifierAndLinkNameTuple

object_reference

See ObjectReference

path_to_object_identifiers

See PathToObjectIdentifiers

policy_attachment

See PolicyAttachment

policy_to_path

See PolicyToPath

rule

See Rule

schema_facet

See SchemaFacet

tag

See Tag

typed_attribute_value_range

See TypedAttributeValueRange

typed_link_attribute_definition

See TypedLinkAttributeDefinition

typed_link_attribute_range

See TypedLinkAttributeRange

typed_link_facet

See TypedLinkFacet

typed_link_facet_attribute_update

See TypedLinkFacetAttributeUpdate

typed_link_schema_and_facet_name

See TypedLinkSchemaAndFacetName

typed_link_specifier

See TypedLinkSpecifier

Structs

AttributeKey

A unique identifier for an attribute.

AttributeKeyAndValue

The combination of an attribute key and an attribute value.

AttributeNameAndValue

Identifies the attribute name and value for a typed link.

BatchAddFacetToObject

Represents the output of a batch add facet to object operation.

BatchAddFacetToObjectResponse

The result of a batch add facet to object operation.

BatchAttachObject

Represents the output of an AttachObject operation.

BatchAttachObjectResponse

Represents the output batch AttachObject response operation.

BatchAttachPolicy

Attaches a policy object to a regular object inside a BatchRead operation. For more information, see AttachPolicy and BatchReadRequest$Operations.

BatchAttachPolicyResponse

Represents the output of an AttachPolicy response

BatchAttachToIndex

Attaches the specified object to the specified index inside a BatchRead operation. For more information, see AttachToIndex and BatchReadRequest$Operations.

BatchAttachToIndexResponse

Represents the output of a AttachToIndex response operation.

BatchAttachTypedLink

Attaches a typed link to a specified source and target object inside a BatchRead operation. For more information, see AttachTypedLink and BatchReadRequest$Operations.

BatchAttachTypedLinkResponse

Represents the output of a AttachTypedLink response operation.

BatchCreateIndex

Creates an index object inside of a BatchRead operation. For more information, see CreateIndex and BatchReadRequest$Operations.

BatchCreateIndexResponse

Represents the output of a CreateIndex response operation.

BatchCreateObject

Represents the output of a CreateObject operation.

BatchCreateObjectResponse

Represents the output of a CreateObject response operation.

BatchDeleteObject

Represents the output of a DeleteObject operation.

BatchDeleteObjectResponse

Represents the output of a DeleteObject response operation.

BatchDetachFromIndex

Detaches the specified object from the specified index inside a BatchRead operation. For more information, see DetachFromIndex and BatchReadRequest$Operations.

BatchDetachFromIndexResponse

Represents the output of a DetachFromIndex response operation.

BatchDetachObject

Represents the output of a DetachObject operation.

BatchDetachObjectResponse

Represents the output of a DetachObject response operation.

BatchDetachPolicy

Detaches the specified policy from the specified directory inside a BatchWrite operation. For more information, see DetachPolicy and BatchWriteRequest$Operations.

BatchDetachPolicyResponse

Represents the output of a DetachPolicy response operation.

BatchDetachTypedLink

Detaches a typed link from a specified source and target object inside a BatchRead operation. For more information, see DetachTypedLink and BatchReadRequest$Operations.

BatchDetachTypedLinkResponse

Represents the output of a DetachTypedLink response operation.

BatchGetLinkAttributes

Retrieves attributes that are associated with a typed link inside a BatchRead operation. For more information, see GetLinkAttributes and BatchReadRequest$Operations.

BatchGetLinkAttributesResponse

Represents the output of a GetLinkAttributes response operation.

BatchGetObjectAttributes

Retrieves attributes within a facet that are associated with an object inside an BatchRead operation. For more information, see GetObjectAttributes and BatchReadRequest$Operations.

BatchGetObjectAttributesResponse

Represents the output of a GetObjectAttributes response operation.

BatchGetObjectInformation

Retrieves metadata about an object inside a BatchRead operation. For more information, see GetObjectInformation and BatchReadRequest$Operations.

BatchGetObjectInformationResponse

Represents the output of a GetObjectInformation response operation.

BatchListAttachedIndices

Lists indices attached to an object inside a BatchRead operation. For more information, see ListAttachedIndices and BatchReadRequest$Operations.

BatchListAttachedIndicesResponse

Represents the output of a ListAttachedIndices response operation.

BatchListIncomingTypedLinks

Returns a paginated list of all the incoming TypedLinkSpecifier information for an object inside a BatchRead operation. For more information, see ListIncomingTypedLinks and BatchReadRequest$Operations.

BatchListIncomingTypedLinksResponse

Represents the output of a ListIncomingTypedLinks response operation.

BatchListIndex

Lists objects attached to the specified index inside a BatchRead operation. For more information, see ListIndex and BatchReadRequest$Operations.

BatchListIndexResponse

Represents the output of a ListIndex response operation.

BatchListObjectAttributes

Represents the output of a ListObjectAttributes operation.

BatchListObjectAttributesResponse

Represents the output of a ListObjectAttributes response operation.

BatchListObjectChildren

Represents the output of a ListObjectChildren operation.

BatchListObjectChildrenResponse

Represents the output of a ListObjectChildren response operation.

BatchListObjectParentPaths

Retrieves all available parent paths for any object type such as node, leaf node, policy node, and index node objects inside a BatchRead operation. For more information, see ListObjectParentPaths and BatchReadRequest$Operations.

BatchListObjectParentPathsResponse

Represents the output of a ListObjectParentPaths response operation.

BatchListObjectParents

Lists parent objects that are associated with a given object in pagination

BatchListObjectParentsResponse

Represents the output of a ListObjectParents response operation.

BatchListObjectPolicies

Returns policies attached to an object in pagination fashion inside a BatchRead operation. For more information, see ListObjectPolicies and BatchReadRequest$Operations.

BatchListObjectPoliciesResponse

Represents the output of a ListObjectPolicies response operation.

BatchListOutgoingTypedLinks

Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object inside a BatchRead operation. For more information, see ListOutgoingTypedLinks and BatchReadRequest$Operations.

BatchListOutgoingTypedLinksResponse

Represents the output of a ListOutgoingTypedLinks response operation.

BatchListPolicyAttachments

Returns all of the ObjectIdentifiers to which a given policy is attached inside a BatchRead operation. For more information, see ListPolicyAttachments and BatchReadRequest$Operations.

BatchListPolicyAttachmentsResponse

Represents the output of a ListPolicyAttachments response operation.

BatchLookupPolicy

Lists all policies from the root of the Directory to the object specified inside a BatchRead operation. For more information, see LookupPolicy and BatchReadRequest$Operations.

BatchLookupPolicyResponse

Represents the output of a LookupPolicy response operation.

BatchReadException

The batch read exception structure, which contains the exception type and

BatchReadOperation

Represents the output of a BatchRead operation.

BatchReadOperationResponse

Represents the output of a BatchRead response operation.

BatchReadSuccessfulResponse

Represents the output of a BatchRead success response operation.

BatchRemoveFacetFromObject

A batch operation to remove a facet from an object.

BatchRemoveFacetFromObjectResponse

An empty result that represents success.

BatchUpdateLinkAttributes

Updates a given typed link’s attributes inside a BatchRead operation. Attributes to be updated must not contribute to the typed link’s identity, as defined by its IdentityAttributeOrder. For more information, see UpdateLinkAttributes and BatchReadRequest$Operations.

BatchUpdateLinkAttributesResponse

Represents the output of a UpdateLinkAttributes response operation.

BatchUpdateObjectAttributes

Represents the output of a BatchUpdate operation.

BatchUpdateObjectAttributesResponse

Represents the output of a BatchUpdate response operation.

BatchWriteOperation

Represents the output of a BatchWrite operation.

BatchWriteOperationResponse

Represents the output of a BatchWrite response operation.

Directory

Directory structure that includes the directory name and directory ARN.

Facet

A structure that contains Name, ARN, Attributes,

FacetAttribute

An attribute that is associated with the Facet.

FacetAttributeDefinition

A facet attribute definition. See Attribute References for more information.

FacetAttributeReference

The facet attribute reference that specifies the attribute definition that contains the

FacetAttributeUpdate

A structure that contains information used to update an attribute.

IndexAttachment

Represents an index and an attached object.

LinkAttributeAction

The action to take on a typed link attribute value. Updates are only supported for attributes which don’t contribute to link identity.

LinkAttributeUpdate

Structure that contains attribute update information.

ObjectAttributeAction

The action to take on the object attribute.

ObjectAttributeRange

A range of attributes.

ObjectAttributeUpdate

Structure that contains attribute update information.

ObjectIdentifierAndLinkNameTuple

A pair of ObjectIdentifier and LinkName.

ObjectReference

The reference that identifies an object.

PathToObjectIdentifiers

Returns the path to the ObjectIdentifiers that is associated with the

PolicyAttachment

Contains the PolicyType, PolicyId, and the ObjectIdentifier to which it is

PolicyToPath

Used when a regular object exists in a Directory and you want to find

Rule

Contains an Amazon Resource Name (ARN) and parameters that are associated with the

SchemaFacet

A facet.

Tag

The tag structure that contains a tag key and value.

TypedAttributeValueRange

A range of attribute values. For more information, see Range Filters.

TypedLinkAttributeDefinition

A typed link attribute definition.

TypedLinkAttributeRange

Identifies the range of attributes that are used by a specified filter.

TypedLinkFacet

Defines the typed links structure and its attributes. To create a typed link facet, use

TypedLinkFacetAttributeUpdate

A typed link facet attribute update.

TypedLinkSchemaAndFacetName

Identifies the schema Amazon Resource Name (ARN) and facet name for the typed

TypedLinkSpecifier

Contains all the information that is used to uniquely identify a typed link. The

Enums

BatchReadExceptionType

BatchWriteExceptionType

ConsistencyLevel

DirectoryState

FacetAttributeType

FacetStyle

ObjectType

RangeMode

RequiredAttributeBehavior

RuleType

TypedAttributeValue

UpdateActionType