Struct aws_sdk_clouddirectory::model::batch_read_successful_response::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for BatchReadSuccessfulResponse

Implementations

impl Builder

pub fn list_object_attributes(
self,
input: BatchListObjectAttributesResponse
) -> Self

Lists all attributes that are associated with an object.

pub fn set_list_object_attributes(
self,
input: Option<BatchListObjectAttributesResponse>
) -> Self

pub fn list_object_children(
self,
input: BatchListObjectChildrenResponse
) -> Self

Returns a paginated list of child objects that are associated with a given object.

pub fn set_list_object_children(
self,
input: Option<BatchListObjectChildrenResponse>
) -> Self

pub fn get_object_information(
self,
input: BatchGetObjectInformationResponse
) -> Self

Retrieves metadata about an object.

pub fn set_get_object_information(
self,
input: Option<BatchGetObjectInformationResponse>
) -> Self

pub fn get_object_attributes(
self,
input: BatchGetObjectAttributesResponse
) -> Self

Retrieves attributes within a facet that are associated with an object.

pub fn set_get_object_attributes(
self,
input: Option<BatchGetObjectAttributesResponse>
) -> Self

pub fn list_attached_indices(
self,
input: BatchListAttachedIndicesResponse
) -> Self

Lists indices attached to an object.

pub fn set_list_attached_indices(
self,
input: Option<BatchListAttachedIndicesResponse>
) -> Self

pub fn list_object_parent_paths(
self,
input: BatchListObjectParentPathsResponse
) -> Self

Retrieves all available parent paths for any object type such as node, leaf node, policy node, and index node objects. For more information about objects, see Directory Structure.

pub fn set_list_object_parent_paths(
self,
input: Option<BatchListObjectParentPathsResponse>
) -> Self

pub fn list_object_policies(
self,
input: BatchListObjectPoliciesResponse
) -> Self

Returns policies attached to an object in pagination fashion.

pub fn set_list_object_policies(
self,
input: Option<BatchListObjectPoliciesResponse>
) -> Self

pub fn list_policy_attachments(
self,
input: BatchListPolicyAttachmentsResponse
) -> Self

Returns all of the ObjectIdentifiers to which a given policy is attached.

pub fn set_list_policy_attachments(
self,
input: Option<BatchListPolicyAttachmentsResponse>
) -> Self

pub fn lookup_policy(self, input: BatchLookupPolicyResponse) -> Self

Lists all policies from the root of the Directory to the object specified. If there are no policies present, an empty list is returned. If policies are present, and if some objects don't have the policies attached, it returns the ObjectIdentifier for such objects. If policies are present, it returns ObjectIdentifier, policyId, and policyType. Paths that don't lead to the root from the target object are ignored. For more information, see Policies.

pub fn set_lookup_policy(self, input: Option<BatchLookupPolicyResponse>) -> Self

pub fn list_index(self, input: BatchListIndexResponse) -> Self

Lists objects attached to the specified index.

pub fn set_list_index(self, input: Option<BatchListIndexResponse>) -> Self

pub fn list_outgoing_typed_links(
self,
input: BatchListOutgoingTypedLinksResponse
) -> Self

Returns a paginated list of all the outgoing TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

pub fn set_list_outgoing_typed_links(
self,
input: Option<BatchListOutgoingTypedLinksResponse>
) -> Self

pub fn list_incoming_typed_links(
self,
input: BatchListIncomingTypedLinksResponse
) -> Self

Returns a paginated list of all the incoming TypedLinkSpecifier information for an object. It also supports filtering by typed link facet and identity attributes. For more information, see Typed Links.

pub fn set_list_incoming_typed_links(
self,
input: Option<BatchListIncomingTypedLinksResponse>
) -> Self

pub fn get_link_attributes(self, input: BatchGetLinkAttributesResponse) -> Self

The list of attributes to retrieve from the typed link.

pub fn set_get_link_attributes(
self,
input: Option<BatchGetLinkAttributesResponse>
) -> Self

pub fn list_object_parents(self, input: BatchListObjectParentsResponse) -> Self

The list of parent objects to retrieve.

pub fn set_list_object_parents(
self,
input: Option<BatchListObjectParentsResponse>
) -> Self

pub fn build(self) -> BatchReadSuccessfulResponse

Consumes the builder and constructs a BatchReadSuccessfulResponse