Module aws_sdk_clouddirectory::operation

Structs

AddFacetToObject

Adds a new Facet to an object. An object can have more than one facet applied on it.

ApplySchema

Copies the input published schema, at the specified version, into the Directory with the same

AttachObject

Attaches an existing object to another object. An object can be accessed in two

AttachPolicy

Attaches a policy object to a regular object. An object can have a limited number of attached

AttachToIndex

Attaches the specified object to the specified index.

AttachTypedLink

Attaches a typed link to a specified source and target object. For more information, see Typed Links.

BatchRead

Performs all the read operations in a batch.

BatchWrite

Performs all the write operations in a batch. Either all the operations succeed or

CreateDirectory

Creates a Directory by copying the published schema into the

CreateFacet

Creates a new Facet in a schema. Facet creation is allowed only

CreateIndex

Creates an index object. See Indexing and search for more information.

CreateObject

Creates an object in a Directory. Additionally attaches the object to

CreateSchema

Creates a new schema in a development state. A schema can exist in three

CreateTypedLinkFacet

Creates a TypedLinkFacet. For more information, see Typed Links.

DeleteDirectory

Deletes a directory. Only disabled directories can be deleted. A deleted directory cannot be undone. Exercise extreme

DeleteFacet

Deletes a given Facet. All attributes and Rules

DeleteObject

Deletes an object and its associated attributes. Only objects with no children and no

DeleteSchema

Deletes a given schema. Schemas in a development and published state can only be deleted.

DeleteTypedLinkFacet

Deletes a TypedLinkFacet. For more information, see Typed Links.

DetachFromIndex

Detaches the specified object from the specified index.

DetachObject

Detaches a given object from the parent object. The object that is to be detached from the

DetachPolicy

Detaches a policy from an object.

DetachTypedLink

Detaches a typed link from a specified source and target object. For more information, see Typed Links.

DisableDirectory

Disables the specified directory. Disabled directories cannot be read or written to.

EnableDirectory

Enables the specified directory. Only disabled directories can be enabled. Once

GetAppliedSchemaVersion

Returns current applied schema version ARN, including the minor version in use.

GetDirectory

Retrieves metadata about a directory.

GetFacet

Gets details of the Facet, such as facet name, attributes, Rules, or ObjectType. You can call this on all kinds of schema

GetLinkAttributes

Retrieves attributes that are associated with a typed link.

GetObjectAttributes

Retrieves attributes within a facet that are associated with an object.

GetObjectInformation

Retrieves metadata about an object.

GetSchemaAsJson

Retrieves a JSON representation of the schema. See JSON Schema Format for more information.

GetTypedLinkFacetInformation

Returns the identity attribute order for a specific TypedLinkFacet. For more information, see Typed Links.

ListAppliedSchemaArns

Lists schema major versions applied to a directory. If SchemaArn is provided, lists the minor version.

ListAttachedIndices

Lists indices attached to the specified object.

ListDevelopmentSchemaArns

Retrieves each Amazon Resource Name (ARN) of schemas in the development

ListDirectories

Lists directories created within an account.

ListFacetAttributes

Retrieves attributes attached to the facet.

ListFacetNames

Retrieves the names of facets that exist in a schema.

ListIncomingTypedLinks

Returns a paginated list of all the incoming TypedLinkSpecifier

ListIndex

Lists objects attached to the specified index.

ListManagedSchemaArns

Lists the major version families of each managed schema. If a major version ARN is provided as SchemaArn, the minor version revisions in that family are listed instead.

ListObjectAttributes

Lists all attributes that are associated with an object.

ListObjectChildren

Returns a paginated list of child objects that are associated with a given

ListObjectParentPaths

Retrieves all available parent paths for any object type such as node, leaf node,

ListObjectParents

Lists parent objects that are associated with a given object in pagination

ListObjectPolicies

Returns policies attached to an object in pagination fashion.

ListOutgoingTypedLinks

Returns a paginated list of all the outgoing TypedLinkSpecifier

ListPolicyAttachments

Returns all of the ObjectIdentifiers to which a given policy is attached.

ListPublishedSchemaArns

Lists the major version families of each published schema. If a major version ARN is provided as SchemaArn, the minor version revisions in that family are listed instead.

ListTagsForResource

Returns tags for a resource. Tagging is currently supported only for directories with a

ListTypedLinkFacetAttributes

Returns a paginated list of all attribute definitions for a particular TypedLinkFacet. For more information, see Typed Links.

ListTypedLinkFacetNames

Returns a paginated list of TypedLink facet names for a particular schema.

LookupPolicy

Lists all policies from the root of the Directory to the object

PublishSchema

Publishes a development schema with a major version and a recommended minor version.

PutSchemaFromJson

Allows a schema to be updated using JSON upload. Only available for development schemas. See JSON Schema Format for more information.

RemoveFacetFromObject

Removes the specified facet from the specified object.

TagResource

An API operation for adding tags to a resource.

UntagResource

An API operation for removing tags from a resource.

UpdateFacet

Does the following:

UpdateLinkAttributes

Updates a given typed link’s attributes. Attributes to be updated must not contribute to the typed link’s identity, as defined by its IdentityAttributeOrder.

UpdateObjectAttributes

Updates a given object's attributes.

UpdateSchema

Updates the schema name with a new name. Only development schema names can be

UpdateTypedLinkFacet

Updates a TypedLinkFacet. For more information, see Typed Links.

UpgradeAppliedSchema

Upgrades a single directory in-place using the PublishedSchemaArn with schema updates found in MinorVersion. Backwards-compatible minor version upgrades are instantaneously available for readers on all objects in the directory. Note: This is a synchronous API call and upgrades only one schema on a given directory per call. To upgrade multiple directories from one schema, you would need to call this API on each directory.

UpgradePublishedSchema

Upgrades a published schema under a new minor version revision using the current contents of DevelopmentSchemaArn.